fix: only display wallet connected chains in related asset selection
